Java doc
I have been having trouble generating the javadoc, I was doing this from
a different computer since my other one is not working. I finally got
the file set up to generate javadoc on this computer. It was already
late last night when I was trying to fix this problem. Please understand
everything includeing the assignment for next week was completed before
the javadoc was created.